+++ src/test/java/by/onliner/footerpages/FT007Test.java
@@ -41,4 +41,29 @@ public class FT007Test {
     public void afterEach() {
         driver.quit();
     }
+
+    public static class FT003Test extends FooterBaseTest {
+
+        @Test
+        public void testWhenUserClickVacanciesThenVacanciesPageOpened() {
+
+            String aVacanciesXPath = "//a[contains(text(),'Вакансии')]";
+            By aVacanciesBy = By.xpath(aVacanciesXPath);
+            WebElement aVacanciesWebElement = driver.findElement(aVacanciesBy);
+
+            aVacanciesWebElement.click();
+
+            String expectedVacanciesURL = "https://blog.onliner.by/vacancy";
+            String actualVacanciesURL = driver.getCurrentUrl();
+            Assertions.assertEquals(expectedVacanciesURL, actualVacanciesURL, "URL страницы вакансий неверный");
+
+            String titleVacanciesTextXPath = "//div[@class='news-header__title']/h1";
+            By titleVacanciesTextBy = By.xpath(titleVacanciesTextXPath);
+            WebElement titleVacanciesTextWebElement = driver.findElement(titleVacanciesTextBy);
+
+            String expectedTitleVacanciesText = "Вакансии";
+            String actualTitleVacanciesText = titleVacanciesTextWebElement.getText();
+            Assertions.assertEquals(expectedTitleVacanciesText, actualTitleVacanciesText);
+        }
+    }
 }
