+++ src/test/java/by/onliner/homepage/HomeBaseTest.java
@@ -5,7 +5,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
 
-public class BaseTest {
+public class HomeBaseTest {
 
     protected WebDriver driver;
     final String BASE_URL = "https://www.onliner.by/";
