+++ src/test/java/by/onliner/homepage/TestST002.java
@@ -7,7 +7,7 @@ import org.openqa.selenium.WebElement;
 import org.openqa.selenium.interactions.Actions;
 import java.util.List;
 
-public class TestST002 extends BaseTest {
+public class TestST002 extends HomeBaseTest {
 
     @Test
     public void testNewsLink() throws InterruptedException {
